ponto_esa_v5/requirements.txt
streamlit
pandas
openpyxl
plotly

# Push Notifications
pywebpush
cryptography

# Scheduler (opcional - para lembretes automáticos)
apscheduler

# API Server (opcional - para endpoints de push)
flask
flask-cors

# PostgreSQL
psycopg2-binary

# Timezone
pytz

# Password hashing
bcrypt

# System monitoring
psutil
